Update Machine Learning Foundation.py
Week 1

This code is to create "Model Representation"
On how to learn to implement the model  𝑓𝑤,𝑏 for linear regression with one variable.

The tools we use is  NumPy and Matplotlib.

As in the lecture, we will use the motivating example of housing price prediction.
This lab will use a simple data set with only two data points - a house with 1000 square feet(sqft) sold for $300,000 and a house with 2000 square feet sold for $500,000. These two points will constitute our data or training set. In this lab, the units of size are 1000 sqft and the units of price are 1000s of dollars.


1. Size (1000 sqft): 1.0 
Price (1000s of dollars): 300

2. Size (1000 sqft): 2.0 
Price (1000s of dollars): 500